// web/src/lib/api/generated/CreateBookRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type CreateBookRequest = {
	title: string;
	author: string | null;
	isbn: string | null;
	language: string | null;
	publisher: string | null;
	series: string | null;
	series_index: number | null;
	page_count: number | null;
	format: string | null;
};