// web/src/lib/api/generated/ProspectiveMetadata.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ProspectiveMetadata = {
	provider: string;
	provider_id: string;
	title: string | null;
	authors: Array<string>;
	description: string | null;
	isbn10: string | null;
	isbn13: string | null;
	page_count: number | null;
	genres: Array<string>;
	cover_url: string | null;
	published_date: string | null;
	publisher: string | null;
	rating: number | null;
	subtitle: string | null;
};